+++ src/main/java/com/yimei/hs/mapper/CangInvoiceDetailMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangInvoiceDetail;
 public interface CangInvoiceDetailMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(CangInvoiceDetail record);
+    long insert(CangInvoiceDetail record);
 
     int insertSelective(CangInvoiceDetail record);
